requirements.txt
numpy
pytorch 
torchvision